requirements.txt
pillow==9.5.0
streamlit==1.31.0